+++ packages/components/.build/stencil/stencil.bindings.vue.ts
@@ -61,3 +61,4 @@ export const VueGenerator = (
     includeUtils: false,
     excludeComponents,
   })
+