+++ packages/components/cypress/support/commands.types.ts
@@ -1,0 +1,7 @@
+/// <reference types="cypress" />
+
+declare namespace Cypress {
+  interface Chainable {
+    visitPage(url: string): void
+  }
+}