+++ packages/components/cypress/support/legacy/pages/accordion.page.ts
@@ -1,0 +1,8 @@
+import { byTestId, AccordionAccessor } from '../../../../../testing/src'
+
+export class AccordionPage {
+  accordion = AccordionAccessor(byTestId('accordion'))
+  open() {
+    cy.visit('/components/bal-accordion')
+  }
+}