+++ packages/components/cypress/support/legacy/pages/datepicker.page.ts
@@ -1,0 +1,8 @@
+import { byTestId, DatePickerAccessor } from '../../../../../testing/src'
+
+export class DatepickerPage {
+  datepicker = DatePickerAccessor(byTestId('datepicker'))
+  open() {
+    cy.visit('/components/form/bal-datepicker')
+  }
+}