+++ packages/components/cypress/support/legacy/pages/input.page.ts
@@ -1,0 +1,8 @@
+import { InputAccessor, byTestId } from '../../../../../testing/src'
+
+export class InputPage {
+  input = InputAccessor(byTestId('input'))
+  open() {
+    cy.visit('/components/form/bal-input')
+  }
+}