+++ packages/components/cypress/support/pages/hint.page.ts
@@ -1,0 +1,11 @@
+import { byTestId } from '../../../../testing/src'
+
+export class HintPage {
+  hint = byTestId('hint')
+  fieldHint = byTestId('field-hint')
+  dataHint = byTestId('data-hint')
+
+  open() {
+    cy.visit('/components/bal-hint')
+  }
+}