+++ packages/components/cypress/support/pages/popover.page.ts
@@ -1,0 +1,8 @@
+import { byTestId } from '../../../../testing/src'
+
+export class PopoverPage {
+  popover = byTestId('popover')
+  open() {
+    cy.visit('/components/bal-popover')
+  }
+}