+++ packages/components/cypress/support/pages/snackbar.page.ts
@@ -1,0 +1,10 @@
+import { byTestId } from '../../../../testing/src'
+
+export class SnackbarPage {
+  snackbar = byTestId('snack')
+  snackbarWarning = byTestId('snack-warning')
+
+  open() {
+    cy.visit('/components/notice/bal-snackbar')
+  }
+}