+++ packages/components/cypress/support/pages/textarea.page.ts
@@ -1,0 +1,10 @@
+import { byTestId } from '../../../../testing/src'
+
+export class TextareaPage {
+  textarea = byTestId('textarea')
+  textareaDisabled = byTestId('textarea-disabled')
+
+  open() {
+    cy.visit('/components/form/bal-textarea')
+  }
+}