+++ packages/components/src/components/bal-tabs/bal-tabs.tsx
@@ -42,6 +42,11 @@ export class Tabs {
   @Prop() action = false
 
   /**
+   * If `true` a light border is shown for the tabs.
+   */
+  @Prop() border = false
+
+  /**
    * Label for the action button
    */
   @Prop() actionLabel = 'Action'
@@ -198,6 +203,7 @@ export class Tabs {
         <Tabs
           value={this.value}
           tabs={this.tabsOptions}
+          border={this.border}
           expanded={this.expanded}
           clickable={this.clickable}
           action={this.action}