+++ packages/components/src/components/bal-tabs/components/tabs.tsx
@@ -5,6 +5,7 @@ import { TabItem } from './tab-item'
 export const TabList: FunctionalComponent<TabProps> = ({
   value,
   expanded,
+  border,
   tabs,
   onSelectTab,
   action,
@@ -13,7 +14,13 @@ export const TabList: FunctionalComponent<TabProps> = ({
   lineWidth,
   lineOffsetLeft,
 }) => (
-  <div class={['tabs', expanded ? 'is-fullwidth' : ''].join(' ')}>
+  <div
+    class={{
+      'tabs': true,
+      'is-fullwidth': expanded,
+      'has-border-bottom-light': border,
+    }}
+  >
     <div class="selected-tab-line" style={{ left: `${lineOffsetLeft || 0}px`, width: `${lineWidth || 0}px` }}></div>
     <ul>
       {tabs.map((tab, index) => (