+++ packages/testing/cypress/support/legacy/pages/list.page.ts
@@ -1,8 +1,0 @@
-import { ListAccessor, byTestId } from '../../../../src'
-
-export class ListPage {
-  list = ListAccessor(byTestId('pagination'))
-  open() {
-    cy.visit('/components/bal-pagination')
-  }
-}