+++ packages/testing/cypress/support/legacy/pages/tile.page.ts
@@ -1,8 +1,0 @@
-import { TileAccessor, byTestId } from '../../../../src'
-
-export class TilePage {
-  tile = TileAccessor(byTestId('card'))
-  open() {
-    cy.visit('/components/bal-card')
-  }
-}