+++ packages/testing/cypress/support/pages/accordion.page.ts
@@ -1,8 +1,0 @@
-import { byTestId } from '../../../src'
-
-export class AccordionPage {
-  accordion = byTestId('accordion')
-  open() {
-    cy.visit('/components/bal-accordion')
-  }
-}